moved index and signup out of tabs
thats it...
