+++ konnekt_app/src/components/auth/SignUp.tsx
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from "react";
 import { View, TextInput, Text, StyleSheet, TouchableOpacity } from "react-native";
 import { useRouter } from 'expo-router';
 
-const API_URL = "http://[YOUR-IP-HERE]:5000/api/auth";  // change ip address before push
+const API_URL = "http://[place-ip-here]:5000/api/auth";  // change ip address before push
 
 export default function SignUp() {
   const router = useRouter();
@@ -127,7 +127,7 @@ const styles = StyleSheet.create({
       backgroundColor: '#F9F9F9',
   },
   button: {
-      backgroundColor: '#4c87df', // Muted lavender for the button
+      backgroundColor: '#4c87df', 
       paddingVertical: 10,
       borderRadius: 5,
       alignItems: 'center',
